csv parser wip; typos, refactor
